+++ scripts/omtk/modules/rigFaceEyes.py
@@ -1,8 +1,8 @@
 import pymel.core as pymel
 import collections
-from omtk.classModule import Module
-from omtk.classModuleFace import ModuleFace
-from omtk.classCtrl import BaseCtrl
+from omtk.core.classModule import Module
+from omtk.core.classModuleFace import ModuleFace
+from omtk.core.classCtrl import BaseCtrl
 from omtk.modules.rigIK import IK
 from omtk.modules.rigFK import FK
 from omtk.libs import libRigging, libCtrlShapes
