+++ scripts/omtk/modules/rigFaceSquint.py
@@ -1,11 +1,11 @@
 import pymel.core as pymel
 import collections
 
-import omtk.classAvar
+from omtk.modules import rigFaceAvar
 from omtk.libs import libCtrlShapes
-from omtk import classModuleFace
+from omtk.core import classModuleFace
 
-class CtrlSquint(omtk.classAvar.BaseCtrlFace):
+class CtrlSquint(rigFaceAvarfrom.BaseCtrlFace):
     def __createNode__(self, **kwargs):
         return libCtrlShapes.create_triangle_upp()
